Change Samples to be an Array instead of Hash
In order to better accommodate timezones, we no longer store data in
Hash as it can change depending on the perspective of the accessor.
There is now a new method #on which takes a DateTime object, which
allows for the specification of a timezone and it includes all samples
from the given input, till 24 hours into the future.